Try to skip grabbing windows SDK if we've never published a build for this branch - to solve the chicken and egg problem between titanium_mobile and titanium_mobile_windows not having builds of the other to use

+++ Jenkinsfile
@@ -115,11 +115,30 @@ timestamps {
 			// Skip the Windows SDK portion if a PR, we don't need it
 			stage('Windows') {
 				if (!isPR) {
-					step([$class: 'CopyArtifact',
-						projectName: "../titanium_mobile_windows/${targetBranch}",
-						selector: [$class: 'StatusBuildSelector', stable: false],
-						filter: 'dist/windows/'])
-					sh 'rm -rf windows; mv dist/windows/ windows/; rm -rf dist'
+					// This may be the very first build on this branch, so there's no windows build to grab yet
+					def isFirstBuildOnBranch = false
+					try {
+						sh 'curl -O http://builds.appcelerator.com.s3.amazonaws.com/mobile/branches.json'
+						if (fileExists('branches.json')) {
+							def contents = readFile('branches.json')
+							if (!contents.startsWith('<?xml')) { // May be an 'Access denied' xml file/response
+								def branchesJSON = jsonParse(contents)
+								isFirstBuildOnBranch = !(branchesJSON['branches'].contains(env.BRANCH_NAME))
+							}
+						}
+					} catch (err) {
+						// ignore? Not able to grab the branches.json, what should we assume? In 99.9% of the cases, it's not a new build
+					}
+
+					if (!isFirstBuildOnBranch) {
+						step([$class: 'CopyArtifact',
+							projectName: "../titanium_mobile_windows/${targetBranch}",
+							selector: [$class: 'StatusBuildSelector', stable: false],
+							filter: 'dist/windows/'])
+						sh 'rm -rf windows; mv dist/windows/ windows/; rm -rf dist'
+					} else {
+						// TODO Just mark the build unstable for now or somehow denote that we're intentionally skipping Windows right now?
+					}
 				} // !isPR
 			} // stage
